requirements.txt
streamlit==1.30.0
mysql-connector-python==8.2.0
pandas==2.1.0
mysqlclient==2.1.1
SQLAlchemy==2.0.20